// src/types.ts
export type GakubuCode =
  | '111' // 法学部
  | '113' // 経済学部
  | '114' // 経営学部
  | '115' // 理工学部
  | '116' // 薬学部
  | '117' // 農学部
  | '118' // 文芸学部
  | '119' // 医学部
  | '11A' // 生物理工学部
  | '11B' // 工学部
  | '11D' // 産業理工学部
  | '11F' // 総合社会学部
  | '11G' // 建築学部
  | '11M' // 国際学部
  | '11N' // 情報学部
  | '11P' // 看護学部
  | '21C' // 総合文化研究科
  | '21G' // 実学社会起業イノベーション学位プログラム
  | '21H' // 建築学研究科
  | '21J' // 情報学研究科
  | '221' // 医学研究科
  | '231' // 法学研究科
  | '232' // 商学研究科
  | '233' // 経済学研究科
  | '236' // 総合理工学研究科
  | '237' // 薬学研究科
  | '238' // 農学研究科
  | '239' // 生物理工学研究科
  | '23B' // 産業理工学研究科
  | '23C' // システム工学研究科
  | '251' // 法科大学院
  | '311' // 短期大学部
  | '411' // 通信教育部
  | 'Z11' // 共通教養
  | 'Z13' // 外国語
  | 'Z15' // 教職課程
  | 'Z17' // 司書課程
  | 'Z19' // 生涯スポーツ
  | 'Z26' // (その他)
  | 'Z28';